[IMPORTANT] Convert the code into C++14 [BUG FOUND] cMapDrawer is wrong

+++ CrossDaRoad-Beta/CrossDaRoad-Beta/cMapLane.cpp
@@ -42,7 +42,7 @@ int cMapLane::GetID() const
 	return nID;
 }
 
-int cMapLane::GetLaneSize() const
+size_t cMapLane::GetLaneSize() const
 {
 	return sLane.size();
 }
@@ -52,7 +52,7 @@ char cMapLane::GetLaneGraphic(int nPos, bool bWrapAroundPosition) const
 	if (bWrapAroundPosition && abs(nPos) >= GetLaneSize()) {
 		nPos %= GetLaneSize(); /// (-size, +size)
 		if (nPos < 0) {
-			nPos += GetLaneSize(); /// [0, +size)
+			nPos += static_cast<int>(GetLaneSize()); /// [0, +size)
 		}
 	}
 	char cGraphic = sLane[nPos];
